xml: migrate to module system
Convert wowless/xml.lua from a standalone require('wowless.xml') file
into wowless/modules/xml.lua, a proper DI module depending on datalua.
Since the DI framework already instantiates one module set per
product, the lang/attributeTypes tables no longer need to lazily
require and cache other products' data by product name; they read
datalua.xmlflat/datalua.stringenums directly, mirroring cstubs.lua.
